Moved files to board folder for better structure
