+++ app/components/board/EditView.js
